base64ct

Module encoding

Base64 encodings

Constants

PAD 🔒

Padding character

Traits

Encoding

Base64 encoding trait.

Functions

decode_padding 🔒

Validate padding is of the expected length compute unpadded length.

decoded_len 🔒

Get the length of the output from decoding the provided unpadded Base64-encoded input.

encoded_len_inner 🔒

is_pad_ct 🔒

Branchless match that a given byte is the PAD character

validate_last_block 🔒

Validate that the last block of the decoded data round-trips back to the encoded data.