Struct base64ct::Encoder

pub struct Encoder<'o, E: Encoding> { /* private fields */ }

Stateful Base64 encoder with support for buffered, incremental encoding.

The E type parameter can be any type which impls Encoding such as Base64 or Base64Unpadded.

Implementations

impl<'o, E: Encoding> Encoder<'o, E>

pub fn new(output: &'o mut [u8]) -> Result<Self, Error>

Create a new encoder which writes output to the given byte slice.

Output constructed using this method is not line-wrapped.

pub fn new_wrapped( output: &'o mut [u8], width: usize, ending: LineEnding ) -> Result<Self, Error>

Create a new encoder which writes line-wrapped output to the given byte slice.

Output will be wrapped at the specified interval, using the provided line ending. Use LineEnding::default() to use the conventional line ending for the target OS.

Minimum allowed line width is 4.

pub fn encode(&mut self, input: &[u8]) -> Result<(), Error>

Encode the provided buffer as Base64, writing it to the output buffer.

Returns

Ok(bytes) if the expected amount of data was read

Err(Error::InvalidLength) if there is insufficient space in the output buffer

pub fn position(&self) -> usize

Get the position inside of the output buffer where the write cursor is currently located.

pub fn finish(self) -> Result<&'o str, Error>

Finish encoding data, returning the resulting Base64 as a str.

pub fn finish_with_remaining(self) -> Result<(&'o str, &'o mut [u8]), Error>

Finish encoding data, returning the resulting Base64 as a str along with the remaining space in the output buffer.