Trait base64ct::Encoding

pub trait Encoding {
    fn decode(src: impl AsRef<[u8]>, dst: &mut [u8]) -> Result<&[u8], Error>;

    fn decode_in_place(buf: &mut [u8]) -> Result<&[u8], InvalidEncodingError>;

    fn encode<'a>(
        src: &[u8], 
        dst: &'a mut [u8]
    ) -> Result<&'a str, InvalidLengthError>;

    fn encoded_len(bytes: &[u8]) -> usize;
}

Base64 encoding trait.

This trait must be imported to make use of any Base64 variant defined in this crate.

Required methods

fn decode(src: impl AsRef<[u8]>, dst: &mut [u8]) -> Result<&[u8], Error>

Decode a Base64 string into the provided destination buffer.

fn decode_in_place(buf: &mut [u8]) -> Result<&[u8], InvalidEncodingError>

Decode a Base64 string in-place.

NOTE: this method does not (yet) validate that padding is well-formed, if the given Base64 encoding is padded.

fn encode<'a>(
src: &[u8],
dst: &'a mut [u8]
) -> Result<&'a str, InvalidLengthError>

Encode the input byte slice as Base64.

Writes the result into the provided destination slice, returning an ASCII-encoded Base64 string value.

fn encoded_len(bytes: &[u8]) -> usize

Get the length of Base64 produced by encoding the given bytes.

WARNING: this function will return 0 for lengths greater than usize::MAX/4!