Struct base64ct::Decoder

pub struct Decoder<'i, E: Encoding> { /* private fields */ }

Stateful Base64 decoder with support for buffered, incremental decoding.

The E type parameter can be any type which impls Encoding such as Base64 or Base64Unpadded.

Implementations

impl<'i, E: Encoding> Decoder<'i, E>

pub fn new(input: &'i [u8]) -> Result<Self, Error>

Create a new decoder for a byte slice containing contiguous (non-newline-delimited) Base64-encoded data.

Returns

Ok(decoder) on success.

Err(Error::InvalidLength) if the input buffer is empty.

pub fn new_wrapped(input: &'i [u8], line_width: usize) -> Result<Self, Error>

Create a new decoder for a byte slice containing Base64 which line wraps at the given line length.

Trailing newlines are not supported and must be removed in advance.

Newlines are handled according to what are roughly RFC7468 conventions:

[parsers] MUST handle different newline conventions

RFC7468 allows any of the following as newlines, and allows a mixture of different types of newlines:

eol        = CRLF / CR / LF

Returns

Ok(decoder) on success.

Err(Error::InvalidLength) if the input buffer is empty or the line width is zero.

pub fn decode<'o>(&mut self, out: &'o mut [u8]) -> Result<&'o [u8], Error>

Fill the provided buffer with data decoded from Base64.

Enough Base64 input data must remain to fill the entire buffer.

Returns

Ok(bytes) if the expected amount of data was read

Err(Error::InvalidLength) if the exact amount of data couldnโt be read

pub fn decode_to_end<'o>( &mut self, buf: &'o mut Vec<u8>, ) -> Result<&'o [u8], Error>

Decode all remaining Base64 data, placing the result into buf.

If successful, this function will return the total number of bytes decoded into buf.

pub fn remaining_len(&self) -> usize

Get the length of the remaining data after Base64 decoding.

Decreases every time data is decoded.

pub fn is_finished(&self) -> bool

Has all of the input data been decoded?